fix: extract pure URL helpers to a leaf module so router.test.ts skips the xterm import chain
router.test.ts imported ./router, which transitively pulls app.tsx →
terminal-client → @xterm/addon-unicode-graphemes; that addon inflates a
compressed Unicode trie at import time and intermittently throws 'Data
error' in CI, killing the suite before any test runs. The three pure
helpers (windowIdToUrlSegment, urlSegmentToWindowId,
validateTerminalSearch) now live in src/lib/router-url.ts (re-exported
by router.tsx unchanged), and the suite — moved beside the leaf as
src/lib/router-url.test.ts per colocation convention — imports zero
xterm code by construction.
